requirements.txt
groq>=0.11.0
ollama>=0.3.3
python-dotenv>=1.0.0
pandas>=2.0.0
streamlit>=1.30.0
pytest>=7.4.0
praw>=7.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
apify-client>=1.7.0
streamlit-extras>=0.4.0
pillow>=10.0.0